Refactored TGenerator and SKTGenerator for less code duplication
